Rename less 'lesskey' and 'README' files.
The `less` utility requires its configuration file to be compiled with
the `lesskey` program. In order to remind the user to do this, the
configuration file was given a deliberately obtrusive name. This method
was *too* obtrusive, especially when cloning this configuration to new
devices; as such, the config files are now hidden. Those who should see
them will, while those who shouldn't won't.