Revamp `nvim` setup and plugins.
Many changes here. They all broadly fall in the category of "making nvim
scale to our current workflow". We want an emacs-like experience for
neovim.

More specifically, we have:

- Re-added `vim-easymotion`. This setting may be *too* aggressive (in
  particular the mappings of `w`, `e`, `b`, etc), but we'll give it
  another shot.
- Swapped the digits and symbol keys; now that we have easymotion, we
  don't really need numbers. This has been done on the keyboard map
  (`langmap` option) level so it works across modes (no bugs, as far as
  I can tell).
- Fixed several small bugs when it comes to clashing plugins (eg.
  we used to get random garbage in our buffers when leader key timed
  out).
- Finally set up some terminal bindings. I never felt the need to use
  the nvim terminal (the rather clunky keybind didn't help) -- but now
  that I think about Justin MK's talk and realize the utility of this
  "nvim as an OS", I think I see the point. Remapped the terminal escape
  code to something nicer, and boy is the copy-paste functionality
  useful. No more messing with PRIMARY vs CLIPBOARD Xorg clipboards.
  Will try to use this more later on.

Note that `lf.nvim` still doesn't work right (it always opens files in a
new window). The README seems to be confusing buffers, windows, and
tabs; we should reevaluate whether we want to use this. We need some
better file browser, that's for sure...