Add `project.nvim` to neovim setup.
`<leader>fP` will find a particular project (as opposed to a file *in* a
project). This will be useful for quickly navigating to other
repositories we are using (instead of using :e and typing in a long file
path...if only vim supported zsh style /u/l/s->/usr/local/share
completion).

In addition, opening a file inside a project will automatically cd
(buffer-local, of course) to the root of the project. This will make
`<leader>ff` more useful. I think. This might not be needed.

In any case, the project navigation should definitely prove useful.
I wonder if there's a way we could make Telescope list all of its
pickers (the old `<leader><leader>` keybind only lists the *built in*
pickers).