Automatically change nvim PWD when `cd`ing in nvim terminal.
When we cd in the nvim terminal, we want the change to propagate to
neovim. This script introduces a bit of delay when `cd`ing in the neovim
terminal, but its utility should redeem it.

God bless this extensibility. Being able to attach to the running neovim
instance from Python is a lifesaver.