Create custom Telescope finder in Lua, disable `autochdir`.
The great refactoring continues. And we've now written our first
extension to Telescope in Lua! The new script, which we can run with
`<leader>fb`, will search for files in the CWD of the current buffer.
If we press enter to select a file, it'll open it. If we use it to
create a file, it'll do that too!

No more need for `:edit yee.txt`; we can use the Telescope finder
instead. Configurability is key.