Print number of buffers in nvim lualine.
I think I understand the new workflow. We've been using many buffers for
one "tab" (small subset of windows). This means heavy use of the buffer
menu, which is always a bit of a drag.

Instead, we should really treat this like `i3`. Heavy use of tabs for
context; windows for site by side views.