Autostart desktop daemons via sway-session.target.
We start our daemons via systemd and dbus. But sway doesn't export
variables (eg: DISPLAY) to systemd and dbus automatically. So far, we
dealt with this by:

1. Manually exporting everything, then restarting all local daemons.
2. Avoiding systemd whatsoever.

Option 1 is ugly, option 2 led to daemon persistence after logout =>
problems when logging in again.

Now, we have a proper systemd target for a sway session. We export the
systemd variables, then activate the target. All of our daemons depend
on the target => get started correctly.