Remove `marks.lua` due to freezes and CPU issues.
When a terminal buffer outputs a lot of text, marks.nvim hangs the
enclosing nvim instance as it tries to draw marks in that buffer. This
is a problem when compiling kernels.

It's also greatly increasing the CPU usage of nvim, and can cause issues
with our CursorHold autocmds. See
https://github.com/chentoast/marks.nvim/issues/40.

When this issue is fixed we may move back. But for now, ditch it.